add home component to handle context use effect, use states
